Add more routes for comments

+++ api/app/Controllers/Http/CommentController.js
@@ -8,6 +8,7 @@ const { unset } = require('lodash');
 
 const NotFoundExceptionResponse = use('App/Responses/NotFoundExceptionResponse');
 const CommentModel = use('App/Models/Comment');
+const PostModel = use('App/Models/Post');
 
 /**
  * Resourceful controller for interacting with comments
@@ -27,18 +28,39 @@ class CommentController {
     }
 
     /**
-   * Show a list of all comments.
-   * GET comments
-   *
-   * @param {object} ctx
-   * @param {Request} ctx.request
-   * @param {Response} ctx.response
-   * @param {View} ctx.view
-   */
+     * Show a list of all comments.
+     * GET comments
+     *
+     * @param {object} ctx
+     * @param {Request} ctx.request
+     * @param {Response} ctx.response
+     * @param {View} ctx.view
+     */
     async index() {
         const comments = await CommentModel.all();
 
         return comments.toJSON().map(comment => this.transform(comment));
+    }
+
+    /**
+     * Show a list of all comments by post.
+     * GET comments
+     *
+     * @param {object} ctx
+     * @param {Request} ctx.request
+     * @param {Response} ctx.response
+     * @param {View} ctx.view
+     */
+    async indexByPost({ params }) {
+        const posts = await PostModel
+        .query()
+        .where('id', params.postId)
+        .with('comments')
+        .fetch();
+
+        const post = posts.first();
+
+        return post.toJSON().comments.map(comment => this.transform(comment));
     }
 
     /**
@@ -62,6 +84,26 @@ class CommentController {
     }
 
     /**
+     * Create/save a new comment.
+     * POST comments
+     *
+     * @param {object} ctx
+     * @param {Request} ctx.request
+     * @param {User} ctx.user
+     */
+    async storeByPost({ request, user, params }) {
+        const commentData = request.only(['message']);
+
+        commentData.post_id = params.postId;
+        commentData.user_id = user.id;
+
+        unset(commentData, 'postId');
+
+        const comment = await CommentModel.create(commentData);
+        return this.transform(comment.toJSON());
+    }
+
+    /**
    * Display a single comment.
    * GET comments/:id
    *
@@ -71,7 +113,7 @@ class CommentController {
     async show({ params, response }) {
         const comment = await CommentModel.find(params.id);
 
-        if (!comment) {
+        if (!comment || params.postId !== comment.post_id) {
             const e = new NotFoundExceptionResponse('comment_not_found');
             return e.send(response);
         }
@@ -87,9 +129,14 @@ class CommentController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-    async update({ params, request }) {
+    async update({ params, request, response }) {
         const comment = await CommentModel.find(params.id);
         const commentData = request.only(['message']);
+
+        if (!comment || params.postId !== comment.post_id) {
+            const e = new NotFoundExceptionResponse('comment_not_found');
+            return e.send(response);
+        }
 
         comment.merge(commentData);
 
@@ -106,8 +153,13 @@ class CommentController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-    async destroy({ params }) {
+    async destroy({ params, response }) {
         const comment = await CommentModel.find(params.id);
+
+        if (!comment || params.postId !== comment.post_id) {
+            const e = new NotFoundExceptionResponse('comment_not_found');
+            return e.send(response);
+        }
 
         await comment.delete();
